Moved format_labels function to plotly.py
